fix: remove remaining int32() casts in sensor examples
Missed two call sites in event-loop and list-sensors that still
cast sensor.Type to int32 for DefaultSensor (which now takes Type).

+++ examples/sensor/event-loop/main.go
@@ -35,7 +35,7 @@ func probeSensor(
 		}
 	}()
 
-	s := mgr.DefaultSensor(int32(sensorType))
+	s := mgr.DefaultSensor(sensorType)
 	name = s.Name()
 	vendor = s.Vendor()
 	if name == "" || vendor == "" {
